Support resolving TCP destination by hostname & IP address for control plane by default, and all workloads if the new flag is set to false.
We introduces a bugfix to the network mapper, that has an impact on the
way we resolve the TCP data the sniffer detects.
We still don't know what would be the impact of this fix (might detect a
lot of new traffic), so we want by default to enable this only for
traffic originated from the control plane, and after we will understand
and test the impact we will release the fix to all kinds of traffic.