Support setting control plane IPv4 CIDR prefix length
The nerwork mapper and the intents-operator need to match IP addresses
to the control plane on some occasions. For this matching, we are using
the CIDR with the prefix length.
On some platform, like GKE, you need to set the prefix length so the
matching would recognize a workload as the control plane correctly, for
example on GKE https://cloud.google.com/kubernetes-engine/docs/best-practices/networking